Move helpers to telemetry
