+++ packages/tests/README.md
@@ -1,25 +1,9 @@
 # Tests
 
 All the workspaces are tested in this workspace.<br/>
-It helps us have a better control over the test specific dependencies, and more importantly, centralize the mockings.<br/>
-Especially useful for having mock projects, built with specific bundlers and run the real thing.<br/>
-Right now we have mock projects for:
+It helps us have a better control over the test specific dependencies.
 
--   [ESBuild](./src/mocks/projects/esbuild)
--   [Webpack 4](./src/mocks/projects/webpack4)
--   [Webpack 5](./src/mocks/projects/webpack5)
-
-## Use a mock project
-
-You should build the project part of your test suite.
-
-```js
-beforeAll(async () => {
-    const output = await exec(`yarn workspace project-{{bundler}} build`);
-
-    // Setup what you need.
-}, 20000);
-```
+Especially useful for having mock projects, built with specific bundlers and run the real thing.
 
 ## Build everything & Run
