+++ packages/tests/src/plugins/telemetry/common/metrics/webpack.test.ts
@@ -12,7 +12,7 @@ import {
     getAssets,
 } from '@dd/telemetry-plugins/common/metrics/webpack';
 import type { Metric } from '@dd/telemetry-plugins/types';
-import { exec, PROJECTS_ROOT } from '@dd/tests/testHelpers';
+import { exec, PROJECTS_ROOT } from '@dd/tests/plugins/telemetry/testHelpers';
 import path from 'path';
 
 describe('Telemetry Webpack Metrics', () => {
