+++ packages/tests/src/plugins/telemetry/esbuild-plugin/modules.test.ts
@@ -4,7 +4,7 @@
 
 import type { LocalModule } from '@dd/core/types';
 import { getModulesResults } from '@dd/telemetry-plugins/common/modules';
-import { mockMetaFile } from '@dd/tests/testHelpers';
+import { mockMetaFile } from '@dd/tests/plugins/telemetry/testHelpers';
 
 describe('Telemetry ESBuild Modules', () => {
     test('It should add module size to the results', () => {
