+++ packages/tests/src/plugins/telemetry/testHelpers.ts
